Removed redundant calls for switching between stages. Created SceneController. Moved FXML files to new directory.
I have removed the redundant calls for the same action in the homepage controller, and added a switchToHomepage in the new SceneController class. I have also moved the FXML files, as i think they were in the wrong directory, if we needed to fetch them. They might still be hard to retrieve, as in the SceneController, there is a "maybe" null-called function.